+++ lib/public/components/qcFlags/qcFlagTypeColoredBadge.js
@@ -12,15 +12,12 @@
  * or submit itself to any jurisdiction.
  */
 
+import { badge } from '../common/badge.js';
+import { h } from '/js/src/index.js';
+
 /**
- * Format QC flag creator name
- * @param {QcFlag} qcFlag QC flag
- * @return {string} name of person/service who created the flag
+ * Create colored badge for QC flag type
+ * @param {QcFlagType} qcFlagType QC flag type
+ * @return {Component} badge
  */
-export const formatQcFlagCreatedBy = ({ origin, createdBy }) => {
-    const defaultName = 'Anonymous';
-    if (origin && (createdBy?.name ?? defaultName) === defaultName) {
-        return origin;
-    }
-    return createdBy?.name || '-';
-};
+export const qcFlagTypeColoredBadge = ({ color, name }) => badge(h('span.flex-row.g1', name), { color });
