+++ lib/public/components/qcFlags/qcFlagTypesSelectionDropdownModel.js
@@ -14,7 +14,7 @@
 import { SelectionDropdownModel } from '../common/selection/dropdown/SelectionDropdownModel.js';
 import { RemoteData } from '/js/src/index.js';
 import { qcFlagTypesProvider } from '../../services/qcFlags/qcFlagTypesProvider.js';
-import { qcFlagTypeColoredBadge } from '../../views/QcFlags/common/qcFlagTypeColoredBadge.js';
+import { qcFlagTypeColoredBadge } from './qcFlagTypeColoredBadge.js';
 
 /**
  * Model storing state of a selection of QC flag types picked from the list of all the existing types
