+++ lib/public/views/QcFlagTypes/ActiveColumns/qcFlagTypesActiveColumns.js
@@ -15,7 +15,7 @@ import { h } from '/js/src/index.js';
 import { formatTimestamp } from '../../../utilities/formatting/formatTimestamp.js';
 import { textFilter } from '../../../components/Filters/common/filters/textFilter.js';
 import { checkboxes } from '../../../components/Filters/common/filters/checkboxFilter.js';
-import { qcFlagTypeColoredBadge } from '../../QcFlags/common/qcFlagTypeColoredBadge.js';
+import { qcFlagTypeColoredBadge } from '../../../components/qcFlags/qcFlagTypeColoredBadge.js';
 
 /**
  * List of active columns for a QC Flag Types table
