+++ lib/public/views/QcFlags/ActiveColumns/qcFlagsActiveColumns.js
@@ -13,8 +13,8 @@
 
 import { h } from '/js/src/index.js';
 import { formatTimestamp } from '../../../utilities/formatting/formatTimestamp.js';
-import { qcFlagTypeColoredBadge } from '../common/qcFlagTypeColoredBadge.js';
-import { formatQcFlagCreatedBy } from '../common/formatQcFlagCreatedBy.js';
+import { qcFlagTypeColoredBadge } from '../../../components/qcFlags/qcFlagTypeColoredBadge.js';
+import { formatQcFlagCreatedBy } from '../format/formatQcFlagCreatedBy.js';
 
 /**
  * Active columns configuration for QC flags table
