+++ lib/public/views/QcFlags/details/qcFlagDetailsConfiguration.js
@@ -14,7 +14,7 @@
 import { h } from '/js/src/index.js';
 import { frontLink } from '../../../components/common/navigation/frontLink.js';
 import { formatTimestamp } from '../../../utilities/formatting/formatTimestamp.js';
-import { qcFlagTypeColoredBadge } from '../common/qcFlagTypeColoredBadge.js';
+import { qcFlagTypeColoredBadge } from '../../../components/qcFlags/qcFlagTypeColoredBadge.js';
 
 /**
  * Display configuration of given QC flag details
