+++ lib/public/views/QcFlags/format/formatQcFlagCreatedBy.js
@@ -12,12 +12,16 @@
  * or submit itself to any jurisdiction.
  */
 
-import { badge } from '../../../components/common/badge.js';
-import { h } from '/js/src/index.js';
-
 /**
- * Create colored badge for QC flag type
- * @param {QcFlagType} qcFlagType QC flag type
- * @return {Component} badge
+ * Format QC flag creator name
+ *
+ * @param {QcFlag} qcFlag QC flag of which creator name should be formatted
+ * @return {string} name of person/service who created the flag
  */
-export const qcFlagTypeColoredBadge = ({ color, name }) => badge(h('span.flex-row.g1', name), { color });
+export const formatQcFlagCreatedBy = ({ origin, createdBy }) => {
+    const defaultName = 'Anonymous';
+    if (origin && (createdBy?.name ?? defaultName) === defaultName) {
+        return origin;
+    }
+    return createdBy?.name || '-';
+};
